chore: move location of v1
